build: remove dist only when it is exited & log out error messages

+++ build.js
@@ -3,7 +3,7 @@ const webpack = require('webpack');
 const config = require('./webpack.config');
 
 // clean the dist folder
-fs.rmdirSync('dist', {recursive : true});
+fs.existsSync('dist') && fs.rmdirSync('dist', {recursive : true});
 
 // build all versions
 webpack(config).run(webpackCallback);
@@ -15,6 +15,7 @@ function log(msg) {
 
 function webpackCallback(err, stats) {
     if (err) {
+        console.log(err);
         process.exit(1);
     }
     log(stats.toString({
